services/dashboard-ui/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
google-cloud-firestore==2.16.0
requests==2.32.3
google-cloud-pubsub==2.23.1